# packages/codegen/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Multi-Tenant Shopify API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/inventory/stats": {
      "get": {
        "tags": ["inventory"],
        "summary": "Get Stats",
        "operationId": "get_stats_api_v1_inventory_stats_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory": {
      "get": {
        "tags": ["inventory"],
        "summary": "List Items",
        "operationId": "list_items_api_v1_inventory_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_InventoryItemResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["inventory"],
        "summary": "Create Item",
        "operationId": "create_item_api_v1_inventory_post",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItemCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{item_id}": {
      "get": {
        "tags": ["inventory"],
        "summary": "Get Item",
        "operationId": "get_item_api_v1_inventory__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Item Id" }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["inventory"],
        "summary": "Update Item",
        "operationId": "update_item_api_v1_inventory__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Item Id" }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItemPatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["inventory"],
        "summary": "Delete Item",
        "operationId": "delete_item_api_v1_inventory__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Item Id" }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryItemCreateInput": {
        "properties": {
          "name": { "type": "string", "minLength": 1, "title": "Name" },
          "sku": { "type": "string", "minLength": 1, "title": "Sku" },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category"
          },
          "supplier": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Supplier"
          },
          "price": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Price"
          },
          "stock": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Stock",
            "default": 0
          }
        },
        "type": "object",
        "required": ["name", "sku"],
        "title": "InventoryItemCreateInput"
      },
      "InventoryItemPatchInput": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "sku": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Sku"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category"
          },
          "supplier": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Supplier"
          },
          "price": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Price"
          },
          "stock": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Stock"
          }
        },
        "type": "object",
        "title": "InventoryItemPatchInput"
      },
      "InventoryItemResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "sku": { "type": "string", "title": "Sku" },
          "name": { "type": "string", "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category"
          },
          "image_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Image Url"
          },
          "status": { "type": "string", "title": "Status" },
          "supplier": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Supplier"
          },
          "total_stock": { "type": "integer", "title": "Total Stock" },
          "total_value": { "type": "integer", "title": "Total Value" },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/InventoryVariantResponse"
            },
            "type": "array",
            "title": "Variants"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "sku",
          "name",
          "status",
          "total_stock",
          "total_value",
          "variants",
          "created_at",
          "updated_at"
        ],
        "title": "InventoryItemResponse"
      },
      "InventoryStatsResponse": {
        "properties": {
          "total_skus": { "type": "integer", "title": "Total Skus" },
          "total_value": { "type": "integer", "title": "Total Value" },
          "low_stock_count": { "type": "integer", "title": "Low Stock Count" },
          "out_of_stock_count": {
            "type": "integer",
            "title": "Out Of Stock Count"
          },
          "total_variants": { "type": "integer", "title": "Total Variants" }
        },
        "type": "object",
        "required": [
          "total_skus",
          "total_value",
          "low_stock_count",
          "out_of_stock_count",
          "total_variants"
        ],
        "title": "InventoryStatsResponse"
      },
      "InventoryVariantResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "item_id": { "type": "string", "format": "uuid", "title": "Item Id" },
          "name": { "type": "string", "title": "Name" },
          "sku": { "type": "string", "title": "Sku" },
          "barcode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Barcode"
          },
          "price": { "type": "integer", "minimum": 0.0, "title": "Price" },
          "cost": { "type": "integer", "minimum": 0.0, "title": "Cost" },
          "stock": { "type": "integer", "title": "Stock" },
          "reorder_point": {
            "type": "integer",
            "title": "Reorder Point",
            "default": 0
          },
          "warehouse": {
            "type": "string",
            "title": "Warehouse",
            "default": "Default"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "name",
          "sku",
          "price",
          "cost",
          "stock",
          "created_at",
          "updated_at"
        ],
        "title": "InventoryVariantResponse"
      },
...
      "PaginatedResponse_InventoryItemResponse_": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/InventoryItemResponse" },
            "type": "array",
            "title": "Data"
          },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" }
        },
        "type": "object",
        "required": ["data", "pagination"],
        "title": "PaginatedResponse[InventoryItemResponse]"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": { "type": "integer", "title": "Page" },
          "page_size": { "type": "integer", "title": "Page Size" },
          "total": { "type": "integer", "title": "Total" },
          "total_pages": { "type": "integer", "title": "Total Pages" }
        },
        "type": "object",
        "required": ["page", "page_size", "total", "total_pages"],
        "title": "PaginationMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/admin/auth/login" }
        }
      }
    }
  }
}